+++ TBSLogistics.ApplicationAPI/Controllers/SFeeByTcommandController.cs
@@ -11,23 +11,26 @@ using TBSLogistics.Model.Model.SFeeByTcommandModel;
 using TBSLogistics.Model.Model.SubFeePriceModel;
 using TBSLogistics.Service.Helpers;
 using TBSLogistics.Service.Panigation;
+using TBSLogistics.Service.Repository.Common;
 using TBSLogistics.Service.Services.ProductServiceManage;
 using TBSLogistics.Service.Services.SFeeByTcommandManage;
 
 
 namespace TBSLogistics.ApplicationAPI.Controllers
 {
-
+    [Authorize]
     [Route("api/[controller]")]
     [ApiController]
     public class SFeeByTcommandController : ControllerBase
     {
         private readonly ISFeeByTcommand _SFeeByTcommand;
         private readonly IPaginationService _uriService;
+        private readonly ICommon _common;
 
 
-        public SFeeByTcommandController(ISFeeByTcommand sFeeByTcommand, IPaginationService uriService)
+        public SFeeByTcommandController(ISFeeByTcommand sFeeByTcommand, IPaginationService uriService,ICommon common)
         {
+            _common= common;
             _SFeeByTcommand = sFeeByTcommand;
             _uriService = uriService;
         }
@@ -35,6 +38,12 @@ namespace TBSLogistics.ApplicationAPI.Controllers
         [Route("[action]")]
         public async Task<IActionResult> CreateSFeeByTCommand(List<CreateSFeeByTCommandRequest> request)
         {
+            var checkPermission = await _common.CheckPermission("F0009");
+            if (checkPermission.isSuccess == false)
+            {
+                return BadRequest(checkPermission.Message);
+            }
+
             var Create = await _SFeeByTcommand.CreateSFeeByTCommand(request);
             if (Create.isSuccess == true)
             {
@@ -50,6 +59,12 @@ namespace TBSLogistics.ApplicationAPI.Controllers
         [Route("[action]")]
         public async Task<IActionResult> DeleteSFeeByTCommand(DeleteSFeeByTCommand Id)
         {
+            var checkPermission = await _common.CheckPermission("F0009");
+            if (checkPermission.isSuccess == false)
+            {
+                return BadRequest(checkPermission.Message);
+            }
+
             var delete = await _SFeeByTcommand.DeleteSFeeByTCommand(Id);
             if (delete.isSuccess == true)
             {
@@ -65,6 +80,12 @@ namespace TBSLogistics.ApplicationAPI.Controllers
         [Route("[action]")]
         public async Task<IActionResult> GetListSubFeeIncurredApprove([FromQuery] PaginationFilter filter)
         {
+            var checkPermission = await _common.CheckPermission("F0008");
+            if (checkPermission.isSuccess == false)
+            {
+                return BadRequest(checkPermission.Message);
+            }
+
             var route = Request.Path.Value;
             var pagedData = await _SFeeByTcommand.GetListSubFeeIncurredApprove(filter);
 
@@ -84,6 +105,12 @@ namespace TBSLogistics.ApplicationAPI.Controllers
         [Route("[action]")]
         public async Task<IActionResult> ApproveSubFeeIncurred(List<ApproveSFeeByTCommand> request)
         {
+            var checkPermission = await _common.CheckPermission("F0010");
+            if (checkPermission.isSuccess == false)
+            {
+                return BadRequest(checkPermission.Message);
+            }
+
             var ApproveSubFeePrice = await _SFeeByTcommand.ApproveSubFeeIncurred(request);
 
             if (ApproveSubFeePrice.isSuccess == true)
@@ -100,6 +127,13 @@ namespace TBSLogistics.ApplicationAPI.Controllers
         [Route("[action]")]
         public async Task<IActionResult> GetListSubFeeIncurredByHandling(int id)
         {
+            var checkPermission = await _common.CheckPermission("F0008");
+            if (checkPermission.isSuccess == false)
+            {
+                return BadRequest(checkPermission.Message);
+            }
+
+
             var list = await _SFeeByTcommand.GetListSubFeeIncurredByHandling(id);
             return Ok(list);
         }
