+++ TBSLogistics.ApplicationAPI/Controllers/SubFeePriceController.cs
@@ -10,6 +10,7 @@ using TBSLogistics.Model.Model.SubFeePriceModel;
 using TBSLogistics.Model.Wrappers;
 using TBSLogistics.Service.Helpers;
 using TBSLogistics.Service.Panigation;
+using TBSLogistics.Service.Repository.Common;
 using TBSLogistics.Service.Services.SubFeePriceManage;
 
 namespace TBSLogistics.ApplicationAPI.Controllers
@@ -21,17 +22,25 @@ namespace TBSLogistics.ApplicationAPI.Controllers
     {
         private readonly ISubFeePrice _subFeePrice;
         private readonly IPaginationService _uriService;
+        private readonly ICommon _common;
 
-        public SubFeePriceController(ISubFeePrice subFeePrice, IPaginationService uriService)
+        public SubFeePriceController(ISubFeePrice subFeePrice, IPaginationService uriService, ICommon common)
         {
             _subFeePrice = subFeePrice;
             _uriService = uriService;
+            _common = common;
         }
 
         [HttpPost]
         [Route("[action]")]
         public async Task<IActionResult> CreateSubFeePrice(CreateSubFeePriceRequest request)
         {
+            var checkPermission = await _common.CheckPermission("D0002");
+            if (checkPermission.isSuccess == false)
+            {
+                return BadRequest(checkPermission.Message);
+            }
+
             var CreateSubFeePrice = await _subFeePrice.CreateSubFeePrice(request);
 
             if (CreateSubFeePrice.isSuccess == true)
@@ -48,6 +57,12 @@ namespace TBSLogistics.ApplicationAPI.Controllers
         [Route("[action]")]
         public async Task<IActionResult> UpdateSubFeePrice(long Id, UpdateSubFeePriceRequest request)
         {
+            var checkPermission = await _common.CheckPermission("D0003");
+            if (checkPermission.isSuccess == false)
+            {
+                return BadRequest(checkPermission.Message);
+            }
+
             var UpdateSubFeePrice = await _subFeePrice.UpdateSubFeePrice(Id, request);
 
             if (UpdateSubFeePrice.isSuccess == true)
@@ -64,6 +79,12 @@ namespace TBSLogistics.ApplicationAPI.Controllers
         [Route("[action]")]
         public async Task<IActionResult> ApproveSubFeePrice(List<ApproveSubFee> request)
         {
+            var checkPermission = await _common.CheckPermission("D0004");
+            if (checkPermission.isSuccess == false)
+            {
+                return BadRequest(checkPermission.Message);
+            }
+
             var ApproveSubFeePrice = await _subFeePrice.ApproveSubFeePrice(request);
 
             if (ApproveSubFeePrice.isSuccess == true)
@@ -80,6 +101,12 @@ namespace TBSLogistics.ApplicationAPI.Controllers
         [Route("[action]")]
         public async Task<IActionResult> GetSubFeePriceById(long id)
         {
+            var checkPermission = await _common.CheckPermission("D0004");
+            if (checkPermission.isSuccess == false)
+            {
+                return BadRequest(checkPermission.Message);
+            }
+
             var sfp = await _subFeePrice.GetSubFeePriceById(id);
             return Ok(sfp);
         }
@@ -88,6 +115,12 @@ namespace TBSLogistics.ApplicationAPI.Controllers
         [Route("[action]")]
         public async Task<IActionResult> GetListSubFeePrice([FromQuery] PaginationFilter filter)
         {
+            var checkPermission = await _common.CheckPermission("D0001");
+            if (checkPermission.isSuccess == false)
+            {
+                return BadRequest(checkPermission.Message);
+            }
+
             var route = Request.Path.Value;
             var pagedData = await _subFeePrice.GetListSubFeePrice(filter);
 
@@ -99,6 +132,7 @@ namespace TBSLogistics.ApplicationAPI.Controllers
         [Route("[action]")]
         public async Task<IActionResult> GetListSubFeeSelect()
         {
+
             var list = await _subFeePrice.GetListSubFeeSelect();
             return Ok(list);
         }
@@ -107,6 +141,12 @@ namespace TBSLogistics.ApplicationAPI.Controllers
         [Route("[action]")]
         public async Task<IActionResult> DisableSubFeePrice(List<long> ids)
         {
+            var checkPermission = await _common.CheckPermission("D0005");
+            if (checkPermission.isSuccess == false)
+            {
+                return BadRequest(checkPermission.Message);
+            }
+
             var DisableSubFeePrice = await _subFeePrice.DisableSubFeePrice(ids);
 
             if (DisableSubFeePrice.isSuccess == true)
@@ -123,6 +163,12 @@ namespace TBSLogistics.ApplicationAPI.Controllers
         [Route("[action]")]
         public async Task<IActionResult> DeleteSubFeePrice(List<long> ids)
         {
+            var checkPermission = await _common.CheckPermission("D0006");
+            if (checkPermission.isSuccess == false)
+            {
+                return BadRequest(checkPermission.Message);
+            }
+
             var DeleteSubFeePrice = await _subFeePrice.DeleteSubFeePrice(ids);
 
             if (DeleteSubFeePrice.isSuccess == true)
